+++ plugins/CuraDrive/src/Settings.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2018 Ultimaker B.V.
 # Cura is released under the terms of the LGPLv3 or higher.
 
-from cura import UltimakerCloudAuthentication
+from cura.UltimakerCloud import UltimakerCloudAuthentication
 
 
 class Settings:
